zalupaspb/web/requirements.txt
Django==4.2.11
django-rest-framework==0.1.0
djangorestframework==3.14.0
channels==4.0.0
psycopg2-binary==2.9.9
redis==5.0.1
pyyaml==6.0.1
python-dotenv==1.0.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
cryptography==42.0.4
gunicorn==21.2.0
drf-yasg==1.21.7
celery==5.3.6 